% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tankr.R
\name{diff_widen}
\alias{diff_widen}
\title{Generate a wide version of ranking comparison.}
\usage{
diff_widen(comparison)
}
\arguments{
\item{comparison}{the long form of the comparions (output from rank_diff_pair)}
}
\value{
A wide dataframe with one column per ranking, one row per measurement
}
\description{
This form is useful when viewing changes in ranking in a table view.
Each rank has it's own column, and each measurement it's own row
}
\examples{
todo
}

